backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic>=2.5.0,<3.0.0
pydantic-settings==2.1.0
python-decouple==3.8
httpx==0.25.2
email-validator==2.1.0
bcrypt==4.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
